Dialogue spacing fix.
The first line of dialogue underneath a character's name should *not*
be preceded by a blank line (in the output).  Subsequent lines of
dialogue should.

My new official recommendation is to use a blank line (in the input)
to separate blocks of text whenever the block type is the same as
the previous one, so that two lines of dialogue would now be separated
by a blank line instead of a second .D macro.

However, with no block type appearing twice in a row in my example,
this information isn't imparted.  This makes it even more urgent for
me to come up with a proper example script, that also runs to at
least two pages, and generally gives an example of everything a user
might conceivably want to do.

For more information about this dialogue spacing issue, see my
(admittedly currently private) Identity comic script repository in
which I try out new formatting ideas before committing them here for
public consumption, or the mailing list thread over at
http://lists.gnu.org/archive/html/groff/2012-06/threads.html .