Using center filled text instead of proper centering.
This is in keeping with the theme of emulating a mechanical typewriter.
The page now has on it essentially a grid or matrix of characters,
just as with a TTY.

Update: OK, upon further investigation, we haven't got the grid
effect yet.  Centered lines can still be half a character out of
alignment with left justified ones.  Still, we've lost the arbitrary
limit of a thousand lines of centered text, however implausible
reaching it might seem, which has to be a good thing.