Start to implement "(MORE)" and "(cont'd)"
This needs several improvements to really work:

* I need to lose the blank line before the "(MORE)"
* I need to move the triggering word onto the next page
* I should work the ".nr"s back in, slightly smarter this time
* I should Include the character's name before "(cont'd)", on the
  same line

This last part may prove especially tricky, as the character's name
isn't stored in a variable yet, and indeed is on the line *after*
the ".C".  Furthermore, they may have a "(V.O.)" or an "(O.S.)" after
their name.  This would simplistically render two sets of brackets,
one after the other, and I'm not yet sure whether that would be
correct or not.